TET-291: divide and conquer parser.test.ts
